backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-groq
langchain-tavily
newsapi-python
newspaper3k
lxml_html_clean
pydantic
tavily-python
python-dotenv
httpx
requests